Change rating column to decimal in reviews table
Updated the 'rating' column in the reviews table migration from tinyInteger to decimal(2,1) to allow ratings with one decimal place, such as 4.5.

+++ backend/database/migrations/2025_04_23_180048_create_reviews_table.php
@@ -12,7 +12,7 @@ class CreateReviewsTable extends Migration
             $table->id(); // Usa 'id' estándar de Laravel para la PK
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->foreignId('property_id')->constrained()->onDelete('cascade');
-            $table->tinyInteger('rating');
+            $table->decimal('rating', 2, 1); // permite valores con un decimal, ej: 4.5
             $table->text('review');
             $table->string('language', 10)->nullable();
             $table->string('author_name')->nullable();
